+++ test/cloudbreak.bash
@@ -37,3 +37,17 @@ EOF
     [[ "$result" == "$expected" ]] || $T_fail "expected=\'$expected\' but actual=\'$result\'"
 }
 
+T_consulRecursorOneValidBridgeAndDockerIpShouldbeExcluded() {
+    result=$(consul-recursors <(cat <<EOF
+nameserver 1.2.3.4
+nameserver 172.17.42.1
+nameserver 192.168.59.103
+nameserver 1.1.1.1
+EOF
+) 172.17.42.1 192.168.59.103)
+
+    local expected=" -recursor 1.2.3.4 -recursor 1.1.1.1"
+    [[ "$result" == "$expected" ]] || $T_fail "expected=\'$expected\' but actual=\'$result\'"
+}
+
+